Add an example/test case for a delayed toggle
This situation resembles the following:
 - A model has a boolean field set to 'true'
 - A toggle is rendered based on that model (with checked = true)
 - The toggle is clicked which caused an update request be sent to the server
 - While there has not yet been a response, the user doesn't get any feedback
 - When a response is received, the model is updated
 - The toggle is re-rendered with checked = false

As the user is left without any feedback for their action, I would like
to try to implement something like this:
 - A toggle is rendered with checked = true
 - When the toggle is clicked, internal state is updated (to false)
 - If the internal state !== the checked prop
   - Set a 'pending' state with a matching CSS class
 - If the internal state === the checked prop
   - Remove the 'pending' state and the matching CSS class

This would allow the style sheet to dictate what the switch looks like
after the user has toggled it and before the checked prop has been
updated.
Examples for this style could be replacing the tick or cross with a
small spinner, pulsing the opacity of the toggle, setting a color while
pending, etc.