Allow a toggle to be in a pending state
This means if a toggle is bound to a `checked` prop and you set the
`allowPending` option, the checked state will still be toggled.
This is assuming the checked prop will eventually be updated to match as
a result of a callback, etc.
In the meantime, while the internal state of the toggle and the
`checked` prop are not the same, a pending css class will be applied.

This functionality is controlled by the `allowPending` option to maintain
a backward compatible behaviour, specifically with the __Controlled
Component without onChange__ scenario, which would normally cause a
read-only toggle.

For a more detailed explanation of why this is useful, please see the
previous commit.