Allow a toggle to be in a pending state
This means if a toggle is bound to a `checked` prop and you set the
`allowPending` option, the checked state will still be toggled.
This is assuming the checked prop will eventually be updated to match as
a result of a callback, etc.
In the meantime, while the internal state of the toggle and the
`checked` prop are not the same, a pending css class will be applied.

This functionality is controlled by the `allowPending` option to maintain
a backward compatible behaviour, specifically with the __Controlled
Component without onChange__ scenario, which would normally cause a
read-only toggle.

For a more detailed explanation of why this is useful, please see the
previous commit.

Stage-2
-------

Because all the props from the `<Toggle>` component are passed to the
underlying `<input>` component, this generates an _unknown prop_ error
because `<input>` doesn't normally take an `allowPending` prop.

Because the `allowPending` prop should not be passed to the `<input>`
component, I had to strip that before passing the props and the spread
operator provides a nice was of doing that if the `stage-2` babel preset
is installed.

However, if you would rather do this without the need to install another
package, instead of this:

```
const { allowPending, ...inputProps } = this.props
```

An alternative could be this:

```
var inputProps = Object.assign({}, this.props);
delete inputProps.allowPending;
```

More info available here:
https://facebook.github.io/react/warnings/unknown-prop.html